fix(asr): address PR review feedback (cleanup, changelog, MEG, module split)
- Remove planning/scaffolding files (docs/asr/ASR_IMPLEMENTATION_PLAN.md, the
  paper4 tutorial notebook, scripts/_fill_aasr_readme_numbers.py, the MATLAB
  rASR shims + ASR_REFERENCE_README) and consolidate the per-PR changelog
  fragments into 36.feature.rst / 36.bugfix.rst.
- utils: drop the UP038 noqa in favour of a PEP 604 union.
- adaptive: expand the module docstring, align the mne import with juggler.py,
  and move the Yule-Walker statistics-filter design into _aasr_filter.py.
- core: support MEG via picks="mag"/"grad"/"meg" with a scale-invariant
  variance check (MEG works like EEG), and split the SPD/Riemannian primitives
  into _spd.py.
