--- tests/parity/matlab_reference/rasr_matlab_shims/grassmannfactory.m
@@ -1,4 +1,0 @@
-function Gr = grassmannfactory(n, k)
-% Minimal shim for rASRMatlab parity generation.
-Gr.n = n;
-Gr.k = k;
